Make sure MacomReader works for files with unicode BOM
The BOM adds a couple of characters to the beginning of the string which
are ignored when python reads the file in non binary mode. That means
that all of our line offsets were a couple of characters off. Fix that
by generating the line offsets with file in binary mode.