Add script to give feedback to teachers
The script is able to use network 3 to find the maximal differences in
the output of the different features the network uses. A user can choose
how many of the top different filters he/she want to be reported.

The script currently only look at 500/700 conv-char-8 features. We
should add support in the script for the last 200 conv-char-8 features
and the 200 conv-char-4 features.