Add values to teacher feedback output
The values are useful to judge which of the outputs the filter actually
looks at. For example consider a filter that reacted to the two strings,

 * "dance", and
 * "penis".

From those two reactions we have no idea whether the filters looks for
"penis" or the filter looks for "dance". When we include the actual
activation,

 * "dance" 4.01, and
 * "penis" 5.01,

we can conclude that the filter in fact looks for "penis".