Add discussion of P_l
The discussion is based on a graph that compares the length based weight
function to the uniform weight function. There was not as much to
discuss as the time based weights as we only had a single text length
based function which we could conclude worked better than the uniform
weights.