Return all output from illustrate_filters
Now that we are looking into teacher feedback scripts we no longer only
need to get the output of single individual filters. Instead we want the
maximum activations of all filters. That will allow us to figure out
what each an every filter is looking at.

That can be used by the teacher feedback script to describe to teachers
what particular filters are looking at. Currently the teacher has to
figure that out by themselves from the raw output of the teacher
feedback script.