Add script to test paragraph finding
Add a script that for each author finds a text with at least 5
paragraphs. Those 5 paragraphs are predicted against the authors other
texts to produce the probability that they are written by the same
author. A negative paragraph is also predicted for each author and the
average of the negatives and positives are written to stdout.

The idea is to use the output to build an argument that our networks are
able to locate the most likely ghost written areas of a text.