Move baseline experiments theory to method
We described a large part of the theory behind our baseline methods in
the experiments section. We defined n-grams there and discussed our
previous project. All that has been moved to different sections. Most of
it are now described in the method section but n-grams are defined
already under notation since they are used in the related works section.