Partition applicability of method in subsections
The section had become very long and hard to read. Split it up into 4
sections with a different focus on how applicable the system will be in
the real world.

Also move part of the end of the section to the top as it fit much
better there.

We have to go through each individual section and make sure that the
introduction in each section makes sense.