Fix errors in backprobagation
Primarily rewrite the start of the section to give a better overview of
gradient descent and what the point is behind explaining
backprobagation. Backprobagation is _not_ the way to update the weight
of the network but rather the way to compute the gradient of the network
that can be used by gradient descent to change the weight of the
network.