Fix makefile error reporting
The utility grep returns an exit status of 0 whenever it finds at least
one line matching a pattern and 1 if it finds no lines matching a
pattern. That is the opposite of what we want. We want the build to be a
failure if grep does find the pattern and not a failure when grep
doesn't find the pattern. Just inverse the exit code of grep to get the
correct behaviour.