Refactor experiments section
Move the description of the prediction system results to under each of
the three networks we present. It makes more sense to have it earlier as
readers will otherwise forget what a specific network is before he/she
is represented with the validation results of the network. We still have
a summarizing view at the end that contains the results for all
networks.